fix(a11y): add <title> to all portal layouts; re-enable document-title axe rule
Issue #189 (axe rule 2 of 3).

Problem: teacher/admin persona Playwright specs disabled the
document-title axe rule because routes rendered without a <title>.
Root layout's metadata.title.default ("StudyBuddy — AI-powered
study material for Grades 5–12") should have cascaded to every
page, but the admin route group's layout was a "use client"
component, which cannot export metadata — so the cascade was
broken for the entire (admin) branch.

Fix, in three parts:

1. (admin)/layout.tsx split: the previous single-file client
   layout is now (a) a server layout that exports
   `metadata: { title: "Admin | StudyBuddy" }` and (b) a new
   AdminAuthShell client component that holds the existing auth
   guard + router logic. Behaviour unchanged; metadata now flows.

2. Explicit `metadata` added to (student), (teacher), (school)
   layouts. All three were already server components, so a
   direct `export const metadata = { title: ... }` works. Adds
   defence-in-depth — route-group-specific titles are clearer
   than relying solely on the root default, and avoids any Next.js
   16 cascade edge case that might surface later.

3. `"document-title"` removed from KNOWN_A11Y_EXCLUSIONS in all
   three persona specs. README updated to reflect the 2 remaining
   exclusions (color-contrast, html-has-lang — the latter shipped
   in a sibling PR against the same umbrella).

No behaviour change for authenticated users: the admin client shell
runs the same useEffect-based redirect, same parseAdminName logic,
same layout shell. Only difference is that the route group's layout
file is now a server component that renders the client shell.

Refs #189
