+++ web/app/(admin)/AdminAuthShell.tsx
@@ -1,0 +1,54 @@
+"use client";
+
+import { useEffect, useRef, useState } from "react";
+import { useRouter } from "next/navigation";
+import { AdminNav } from "@/components/layout/AdminNav";
+import { QueryProvider } from "@/lib/providers/QueryProvider";
+import { PortalHeader } from "@/components/layout/PortalHeader";
+import { PortalFooter } from "@/components/layout/PortalFooter";
+
+function parseAdminName(token: string): string | undefined {
+  try {
+    const payload = JSON.parse(
+      atob(token.split(".")[1].replace(/-/g, "+").replace(/_/g, "/")),
+    );
+    // Prefer name > email (show the local-part before @ for brevity) > admin_id UUID as last resort.
+    if (payload.name) return payload.name;
+    if (payload.email) return String(payload.email).split("@")[0];
+    return payload.sub ?? payload.admin_id ?? undefined;
+  } catch {
+    return undefined;
+  }
+}
+
+export function AdminAuthShell({ children }: { children: React.ReactNode }) {
+  const router = useRouter();
+  const [userName, setUserName] = useState<string | undefined>(undefined);
+  const redirectedRef = useRef(false);
+
+  useEffect(() => {
+    if (redirectedRef.current) return;
+    const token = localStorage.getItem("sb_admin_token");
+    if (!token) {
+      redirectedRef.current = true;
+      router.replace("/admin/login");
+    } else {
+      setUserName(parseAdminName(token));
+    }
+  }, [router]);
+
+  return (
+    <QueryProvider>
+      <div className="flex min-h-screen bg-gray-50">
+        <AdminNav />
+        <div className="flex flex-1 flex-col overflow-auto">
+          <PortalHeader portal="admin" userName={userName} />
+          <main id="main-content" className="flex-1">
+            {children}
+          </main>
+          <PortalFooter />
+        </div>
+      </div>
+    </QueryProvider>
+  );
+}
