+++ web/app/(admin)/layout.tsx
@@ -1,54 +1,12 @@
-"use client";
+import type { Metadata } from "next";
+import { AdminAuthShell } from "./AdminAuthShell";
 
-import { useEffect, useRef, useState } from "react";
-import { useRouter } from "next/navigation";
-import { AdminNav } from "@/components/layout/AdminNav";
-import { QueryProvider } from "@/lib/providers/QueryProvider";
-import { PortalHeader } from "@/components/layout/PortalHeader";
-import { PortalFooter } from "@/components/layout/PortalFooter";
-
-function parseAdminName(token: string): string | undefined {
-  try {
-    const payload = JSON.parse(
-      atob(token.split(".")[1].replace(/-/g, "+").replace(/_/g, "/")),
-    );
-    // Prefer name > email (show the local-part before @ for brevity) > admin_id UUID as last resort.
-    if (payload.name) return payload.name;
-    if (payload.email) return String(payload.email).split("@")[0];
-    return payload.sub ?? payload.admin_id ?? undefined;
-  } catch {
-    return undefined;
-  }
-}
+// Server layout (no `"use client"`) so the Metadata API can set <title>.
+// All interactive/auth-guard logic lives in the AdminAuthShell client component.
+export const metadata: Metadata = {
+  title: "Admin | StudyBuddy",
+};
 
 export default function AdminLayout({ children }: { children: React.ReactNode }) {
-  const router = useRouter();
-  const [userName, setUserName] = useState<string | undefined>(undefined);
-  const redirectedRef = useRef(false);
-
-  useEffect(() => {
-    if (redirectedRef.current) return;
-    const token = localStorage.getItem("sb_admin_token");
-    if (!token) {
-      redirectedRef.current = true;
-      router.replace("/admin/login");
-    } else {
-      setUserName(parseAdminName(token));
-    }
-  }, [router]);
-
-  return (
-    <QueryProvider>
-      <div className="flex min-h-screen bg-gray-50">
-        <AdminNav />
-        <div className="flex flex-1 flex-col overflow-auto">
-          <PortalHeader portal="admin" userName={userName} />
-          <main id="main-content" className="flex-1">
-            {children}
-          </main>
-          <PortalFooter />
-        </div>
-      </div>
-    </QueryProvider>
-  );
+  return <AdminAuthShell>{children}</AdminAuthShell>;
 }
