+++ web/tests/e2e/personas/admin-accessibility.spec.ts
@@ -20,7 +20,7 @@ import { checkA11y } from "../helpers/axe";
 // Axe rules disabled for this persona. Tracked in GitHub issue #189
 // (umbrella Epic 9 accessibility audit). Do NOT extend without filing
 // a corresponding issue first.
-const KNOWN_A11Y_EXCLUSIONS = ["color-contrast", "document-title"] as const;
+const KNOWN_A11Y_EXCLUSIONS = ["color-contrast"] as const;
 import { makeAdminToken } from "../helpers/tokens";
 
 // ---------------------------------------------------------------------------
